fix bug

+++ app/src/main/java/com/vietinterview/getbee/api/response/detailjob/DetailJobResponse.java
@@ -34,16 +34,16 @@ public class DetailJobResponse implements Parcelable {
     private Integer quantity;
     @SerializedName("fromSalary")
     @Expose
-    private Integer fromSalary;
+    private Long fromSalary;
     @SerializedName("toSalary")
     @Expose
-    private Integer toSalary;
+    private Long toSalary;
     @SerializedName("currency")
     @Expose
     private Integer currency;
     @SerializedName("fee")
     @Expose
-    private Integer fee;
+    private Long fee;
     @SerializedName("jobDescription")
     @Expose
     private String jobDescription;
@@ -144,12 +144,12 @@ public class DetailJobResponse implements Parcelable {
         if (in.readByte() == 0) {
             fromSalary = null;
         } else {
-            fromSalary = in.readInt();
+            fromSalary = in.readLong();
         }
         if (in.readByte() == 0) {
             toSalary = null;
         } else {
-            toSalary = in.readInt();
+            toSalary = in.readLong();
         }
         if (in.readByte() == 0) {
             currency = null;
@@ -159,7 +159,7 @@ public class DetailJobResponse implements Parcelable {
         if (in.readByte() == 0) {
             fee = null;
         } else {
-            fee = in.readInt();
+            fee = in.readLong();
         }
         jobDescription = in.readString();
         submitDate = in.readString();
@@ -325,19 +325,19 @@ public class DetailJobResponse implements Parcelable {
         this.quantity = quantity;
     }
 
-    public Integer getFromSalary() {
+    public Long getFromSalary() {
         return fromSalary;
     }
 
-    public void setFromSalary(Integer fromSalary) {
+    public void setFromSalary(Long fromSalary) {
         this.fromSalary = fromSalary;
     }
 
-    public Integer getToSalary() {
+    public Long getToSalary() {
         return toSalary;
     }
 
-    public void setToSalary(Integer toSalary) {
+    public void setToSalary(Long toSalary) {
         this.toSalary = toSalary;
     }
 
@@ -349,11 +349,11 @@ public class DetailJobResponse implements Parcelable {
         this.currency = currency;
     }
 
-    public Integer getFee() {
+    public Long getFee() {
         return fee;
     }
 
-    public void setFee(Integer fee) {
+    public void setFee(Long fee) {
         this.fee = fee;
     }
 
@@ -560,13 +560,13 @@ public class DetailJobResponse implements Parcelable {
             parcel.writeByte((byte) 0);
         } else {
             parcel.writeByte((byte) 1);
-            parcel.writeInt(fromSalary);
+            parcel.writeLong(fromSalary);
         }
         if (toSalary == null) {
             parcel.writeByte((byte) 0);
         } else {
             parcel.writeByte((byte) 1);
-            parcel.writeInt(toSalary);
+            parcel.writeLong(toSalary);
         }
         if (currency == null) {
             parcel.writeByte((byte) 0);
@@ -578,7 +578,7 @@ public class DetailJobResponse implements Parcelable {
             parcel.writeByte((byte) 0);
         } else {
             parcel.writeByte((byte) 1);
-            parcel.writeInt(fee);
+            parcel.writeLong(fee);
         }
         parcel.writeString(jobDescription);
         parcel.writeString(submitDate);
